[FE]REFACTOR: Mapinfo 코드 리팩토링_MapGridSilder 경로 수정
