+++ frontend/src/components/MapInfo/MapGridSilder.container.tsx
